CAMBIOS

+++ plugin/src/main/java/io/github/wickeddroid/plugin/world/WorldGenerator.java
@@ -7,6 +7,7 @@ import org.bukkit.*;
 import team.unnamed.inject.Inject;
 import team.unnamed.inject.Singleton;
 
+import java.io.File;
 import java.io.IOException;
 
 @Singleton
@@ -33,7 +34,7 @@ public class WorldGenerator {
       return;
     }
 
-    this.setupWorldBorder(this.world, this.worlds.worldBorder());
+    this.setupWorldBorder(this.world, this.worlds.border().worldBorder());
     this.applySettings(this.world);
   }
 
@@ -43,9 +44,26 @@ public class WorldGenerator {
       return;
     }
 
+    if(!worlds.removeWorld()) { return; }
+
     Bukkit.unloadWorld(world, false);
 
     FileUtils.deleteDirectory(world.getWorldFolder());
+  }
+
+  public void removeLobbyData() {
+    worlds.blacklist().forEach(w -> {
+      var world = Bukkit.getWorld(w);
+      if(w == null) { return; }
+
+      try {
+        FileUtils.deleteDirectory(new File(world.getWorldFolder().getAbsolutePath()+"/stats/"));
+        FileUtils.deleteDirectory(new File(world.getWorldFolder().getAbsolutePath()+"/advancements/"));
+        FileUtils.deleteDirectory(new File(world.getWorldFolder().getAbsolutePath()+"/playerdata/"));
+      } catch (IOException e) {
+        throw new RuntimeException(e);
+      }
+    });
   }
 
   public void applySettings(final World world) {
